Switch to the 'rustc' manifest and build repack with install.sh.
The channel-rustc-* manifests list bare toolchains and std library
builds separately. Prefer this to the bundled channel-rust-*
manifest since it's easier to add additional targets this way.

For each target platform, unpack the compiler and run the included
install.sh to install it in our generic 'rustc' subtree for use
in integration builds. Then do the same with the corresponding
std library package.

Also set -e so the script terminates when there's a problem.